requirements.txt
PyYAML>=6.0
paho-mqtt>=1.6.0
pysnmp>=6.0
pymodbus>=3.0.0
RPi.GPIO>=0.7.0
aiohttp>=3.8.0
requests>=2.28.0